requirements.txt
boto3
psycopg2-binary
pydantic[email]
requests
python-dotenv
